Add worker support for daily indexed collections
Prior to this commit, workers would dump all collected
tweet data into a single, unindexed, monolithic database
collection. Although simple in principle, this storage
strategy could be very costly in the long run. In order to
resolve this issue, this patch adds support for daily
database collections.

In addition to daily segmentation, this patch allows the
async workers to build indexes on the full text, hashtags,
and user mentions found in the collected tweet data.